+++ src/main/java/com/example/kakaoshop/product/Product.java
@@ -13,7 +13,7 @@ public class Product {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     @Column(length = 100, nullable = false)
     private String productName;
     @Column(length = 1000, nullable = false)
@@ -24,7 +24,7 @@ public class Product {
 
 
     @Builder
-    public Product(int id, String productName, String description, String image, int price) {
+    public Product(Long id, String productName, String description, String image, int price) {
         this.id = id;
         this.productName = productName;
         this.description = description;
